(#3) Implemented a mode switcher in the history activity.
The issue originally wanted a separate menu item on the main screen, but trying to have both that AND a mode switcher turned out to have weird side effects, because of a battle between the "last intent" and the saved window state.

Since I'm the one who wrote the issue in the first place, I accept the developer's compromise on this implementation. ^_^